+++ nautobot_circuit_maintenance/__init__.py
@@ -1,6 +1,8 @@
 """Init for Circuit Maintenance plugin."""
-__version__ = "0.1.1"
+__version__ = "0.1.2"
+from django.conf import settings
 from django.db.models.signals import post_migrate
+from django.utils.text import slugify
 from nautobot.extras.plugins import PluginConfig
 
 
@@ -23,6 +25,29 @@ def custom_field_extension(sender, **kwargs):  # pylint: disable=unused-argument
         field.content_types.set([ContentType.objects.get_for_model(Provider)])
 
 
+def import_notification_sources(sender, **kwargs):  # pylint: disable=unused-argument
+    """Import Notification Sources from Nautobot_configuration.py.
+
+    This is a temporary solution until a secrets backend is implemented.
+    For now, we create the Notification Sources in the DB but the secrets are fetched via ENV.
+    """
+    # pylint: disable=import-outside-toplevel
+
+    from nautobot_circuit_maintenance.models import NotificationSource
+
+    desired_notification_sources_names = []
+    for notification_source in settings.PLUGINS_CONFIG.get("nautobot_circuit_maintenance", {}).get(
+        "notification_sources", []
+    ):
+        NotificationSource.objects.get_or_create(
+            name=notification_source["name"], slug=slugify(notification_source["name"])
+        )
+        desired_notification_sources_names.append(notification_source["name"])
+
+    # We remove old Notification Sources that could be in Nautobot but removed from configuration
+    NotificationSource.objects.exclude(name__in=desired_notification_sources_names).delete()
+
+
 class CircuitMaintenanceConfig(PluginConfig):
     """Plugin configuration for the Circuit Maintenance plugin."""
 
@@ -30,7 +55,8 @@ class CircuitMaintenanceConfig(PluginConfig):
     verbose_name = "Circuit Maintenance"
     version = __version__
     author = "Network to Code, LLC"
-    description = ""
+    author_email = "opensource@networktocode.com"
+    description = "Automatically handle network circuit maintenance notifications."
     base_url = "circuit-maintenance"
     min_version = "1.0.0-beta.4"
     max_version = "1.999"
@@ -41,6 +67,7 @@ class CircuitMaintenanceConfig(PluginConfig):
     def ready(self):
         super().ready()
         post_migrate.connect(custom_field_extension, sender=self)
+        post_migrate.connect(import_notification_sources, sender=self)
 
 
 config = CircuitMaintenanceConfig  # pylint:disable=invalid-name