+++ nautobot_circuit_maintenance/handle_notifications/handler.py
@@ -14,7 +14,7 @@ from nautobot_circuit_maintenance.models import (
     RawNotification,
     NotificationSource,
 )
-from .email_helper import get_notifications_from_email
+from .sources import get_notifications
 
 
 # pylint: disable=broad-except
@@ -216,7 +216,7 @@ class HandleCircuitMaintenanceNotifications(Job):
     class Meta:
         """Meta object boilerplate for HandleParsedNotifications."""
 
-        name = "Handle Circuit Mainentance Notifications"
+        name = "Handle Circuit Maintenance Notifications"
         description = (
             "Retrieve via email latest Circuit Maintenance Notifications and create or update them accrodingly"
         )
@@ -239,9 +239,11 @@ class HandleCircuitMaintenanceNotifications(Job):
             last_time_processed = None
 
         try:
-            notifications = get_notifications_from_email(
+            # TODO: get_notifications should be replaced by get_notifications_from_source when new types
+            # are supported
+            notifications = get_notifications(
                 logger=self,
-                email_boxes=notification_sources,
+                notification_sources=notification_sources,
                 since=last_time_processed,
             )
             if not notifications: