+++ CHANGELOG.md
@@ -1,12 +1,22 @@
 # Changelog
 
+## v0.1.2
+
+### Fixed
+
+- #11 - Fix images links from PyPI
+
+### Changed
+
+- #13 - Move Notification Source secrets to configuration.py and refactor class to get ready for new integrations
+
 ## v0.1.1 - 2021-05-14
 
 ### Fixed
 
-- #11 - fix NotificationSource update with previous password
-- #11 - fix Custom Validator update
-- #11 - Rename EmailSettingsServer model to NotificationSource
+- #8 - fix NotificationSource update with previous password
+- #8 - fix Custom Validator update
+- #8 - Rename EmailSettingsServer model to NotificationSource
 
 ## v0.1.0 - 2021-05-03